Base ggproto classes for ggbumpribbon

These ggproto objects implement the statistical transformations for bump ribbon and line charts. They are exported for extensibility but should typically be used through geom_bump_ribbon() and geom_bump_line().

See also